refactor: change radio buttons color implementation
