+++ angular-site/src/app/form-controls/radio-button/components/radio-buttons-disabled/radio-buttons-disabled.component.html
@@ -3,16 +3,16 @@
     Basic radio disabled
   </a>
 </h2>
-<mat-radio-group [color]="colorTheme" disabled id="basic-radio-group-disabled" aria-label="Select an option">
-  <mat-radio-button id="{{'radio-option-one-disabled'}}" value="1">Option 1</mat-radio-button>
-  <mat-radio-button id="{{'radio-option-two-disabled'}}" value="2">Option 2</mat-radio-button>
+<mat-radio-group disabled id="basic-radio-group-disabled" aria-label="Select an option">
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-one-disabled'}}" value="1">Option 1</mat-radio-button>
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-two-disabled'}}" value="2">Option 2</mat-radio-button>
 </mat-radio-group>
 <br/>
 <br/>
-<mat-radio-group [color]="colorTheme" id="basic-radio-group-switch-disable" aria-label="Select an option">
-  <mat-radio-button id="{{'radio-option-one-switch-disabled'}}" value="1">Option 1</mat-radio-button>
-  <mat-radio-button [disabled]="isDisabled" id="{{'radio-option-two-switch-disabled'}}" value="2">Option 2</mat-radio-button>
-  <mat-radio-button [disabled]="isDisabled" id="{{'radio-option-three-switch-disabled'}}" value="3">Option 3</mat-radio-button>
+<mat-radio-group id="basic-radio-group-switch-disable" aria-label="Select an option">
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-one-switch-disabled'}}" value="1">Option 1</mat-radio-button>
+  <mat-radio-button [color]="colorTheme" [disabled]="isDisabled" id="{{'radio-option-two-switch-disabled'}}" value="2">Option 2</mat-radio-button>
+  <mat-radio-button [color]="colorTheme" [disabled]="isDisabled" id="{{'radio-option-three-switch-disabled'}}" value="3">Option 3</mat-radio-button>
 </mat-radio-group>
 <br/>
 <br/>
