+++ angular-site/src/app/form-controls/radio-button/components/radio-buttons-label-position/radio-buttons-label-position.component.html
@@ -3,16 +3,16 @@
     Basic radio label position before
   </a>
 </h2>
-<mat-radio-group [color]="colorTheme" labelPosition="before" id="basic-radio-group-label-position" aria-label="Select an option">
-  <mat-radio-button id="{{'radio-option-one-label-position'}}" value="1">Radio group label position before 1</mat-radio-button>
+<mat-radio-group labelPosition="before" id="basic-radio-group-label-position" aria-label="Select an option">
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-one-label-position'}}" value="1">Radio group label position before 1</mat-radio-button>
   <br/>
-  <mat-radio-button id="{{'radio-option-two-label-position'}}" value="2">Radio group label position before 2</mat-radio-button>
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-two-label-position'}}" value="2">Radio group label position before 2</mat-radio-button>
 </mat-radio-group>
 <br/>
 <br/>
-<mat-radio-group [color]="colorTheme" id="basic-radio-group" aria-label="Select an option">
-  <mat-radio-button labelPosition="before" id="{{'radio-option-one-label-position-default'}}" value="1">Radio label position before 1</mat-radio-button>
+<mat-radio-group id="basic-radio-group" aria-label="Select an option">
+  <mat-radio-button [color]="colorTheme" labelPosition="before" id="{{'radio-option-one-label-position-default'}}" value="1">Radio label position before 1</mat-radio-button>
   <br/>
-  <mat-radio-button id="{{'radio-option-two-label-position-default'}}" value="2">Radio label position default 2</mat-radio-button>
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-two-label-position-default'}}" value="2">Radio label position default 2</mat-radio-button>
 </mat-radio-group>
