+++ angular-site/src/app/form-controls/radio-button/components/radio-buttons-overview/radio-overview-example.html
@@ -1,0 +1,9 @@
+<h2 class="example-h2">
+  <a href="https://material.angular.io/components/radio/overview">
+    Basic radios
+  </a>
+</h2>
+<mat-radio-group id="basic-radio-group" aria-label="Select an option">
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-one'}}" value="1">Option 1</mat-radio-button>
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-two'}}" value="2">Option 2</mat-radio-button>
+</mat-radio-group>
