+++ angular-site/src/app/form-controls/radio-button/components/radio-buttons-ripple-disabled/radio-buttons-ripple-disabled.component.html
@@ -3,10 +3,10 @@
     Basic radio ripple disabled
   </a>
 </h2>
-<mat-radio-group [color]="colorTheme" id="basic-radio-group-ripple-disabled" aria-label="Select an option">
-  <mat-radio-button disableRipple="true" id="{{'radio-option-ripple-disabled'}}" value="1">Radio button ripple disabled</mat-radio-button>
+<mat-radio-group id="basic-radio-group-ripple-disabled" aria-label="Select an option">
+  <mat-radio-button [color]="colorTheme" disableRipple="true" id="{{'radio-option-ripple-disabled'}}" value="1">Radio button ripple disabled</mat-radio-button>
   <br/>
-  <mat-radio-button id="{{'radio-option-ripple-enabled'}}" value="2">Radio button ripple enabled</mat-radio-button>
+  <mat-radio-button [color]="colorTheme" id="{{'radio-option-ripple-enabled'}}" value="2">Radio button ripple enabled</mat-radio-button>
 </mat-radio-group>
