--- angular-site/src/app/form-controls/radio-button/components/radio-overview-example.html
@@ -1,9 +1,0 @@
-<h2 class="example-h2">
-  <a href="https://material.angular.io/components/radio/overview">
-    Basic radios
-  </a>
-</h2>
-<mat-radio-group [color]="colorTheme" id="basic-radio-group" aria-label="Select an option">
-  <mat-radio-button id="{{'radio-option-one'}}" value="1">Option 1</mat-radio-button>
-  <mat-radio-button id="{{'radio-option-two'}}" value="2">Option 2</mat-radio-button>
-</mat-radio-group>
