+++ angular-site/src/app/form-controls/radio-button/radio-buttons.component.html
@@ -1,13 +1,13 @@
 <section class="example-section">
   <label class="example-margin">Color:</label>
-  <mat-radio-group id="radio-buttons-color_buttons" [(ngModel)]="colorTheme" [color]="colorTheme">
-    <mat-radio-button id="radio-buttons_dialog_color_radio" class="example-margin" value="primary">
+  <mat-radio-group id="radio-buttons-color_buttons" [(ngModel)]="colorTheme">
+    <mat-radio-button [color]="colorTheme" id="radio-buttons_dialog_color_radio" class="example-margin" value="primary">
       primary
     </mat-radio-button>
-    <mat-radio-button id="radio-buttons_first-tabbable_color_radio" class="example-margin" value="accent">
+    <mat-radio-button [color]="colorTheme" id="radio-buttons_first-tabbable_color_radio" class="example-margin" value="accent">
       accent
     </mat-radio-button>
-    <mat-radio-button id="radio-buttons_first-heading_color_radio" class="example-margin" value="warn">
+    <mat-radio-button [color]="colorTheme" id="radio-buttons_first-heading_color_radio" class="example-margin" value="warn">
       warn
     </mat-radio-button>
   </mat-radio-group>
